pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dharma-shield-env"
version = "1.0.0"
description = "Policy-driven trust and safety OpenEnv benchmark environment."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn==0.27.0",
  "pydantic==2.6.0",
  "httpx==0.27.0",
  "openai==1.30.0",
  "python-dotenv==1.0.0",
  "pyyaml==6.0.2",
  "openenv-core>=0.2.0",
]

[project.scripts]
server = "server.app:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
